Only send keystrokes once all keys are lifted.
This addresses a problem where the keys are sent prior to the user lifting the windows key. I had 123 as the keys to sent and, when hitting ctrl-z, the result was Win+1, Win+2, Win+3. Now, the keys are only sent after we detect no keys being pressed.

I upgraded the project to 4.6.1 to leverage async.